Move non-fungible with remote type to separate scenario
