+++ transaction-scenarios/generated-examples/kv_store_with_remote_type/001--kv-store-with-remote-type-cdd653eb8baac514b7d7c511a1cac390480b3c92e2eb48738c0e78421da771d2.blob
