+++ transaction-scenarios/generated-examples/kv_store_with_remote_type/001--kv-store-with-remote-type.rtm
