+++ transaction-scenarios/generated-examples/non_fungible_resource_with_remote_type/001--non-fungible-resource-with-remote-type-registration-6ea1d7af76b57f04312dc7c89e9bffe9e2ac47b880f43b01b69b175f3de9fbed.blob
