+++ transaction-scenarios/src/scenarios/all_scenarios.rs
@@ -37,7 +37,10 @@ impl Iterator for AllScenariosIterator {
                 global_n_owned::GlobalNOwnedScenarioCreator::create(core)
             })),
             8 => Some(Box::new(|core| {
-                kv_store::KVStoreScenarioCreator::create(core)
+                non_fungible_resource_with_remote_type::NonFungibleResourceWithRemoteTypeScenarioCreator::create(core)
+            })),
+            9 => Some(Box::new(|core| {
+                kv_store_with_remote_type::KVStoreScenarioCreator::create(core)
             })),
             _ => None,
         }
