+++ transaction-scenarios/src/scenarios/mod.rs
@@ -4,9 +4,10 @@ pub mod account_authorized_depositors;
 mod all_scenarios;
 pub mod fungible_resource;
 pub mod global_n_owned;
-pub mod kv_store;
+pub mod kv_store_with_remote_type;
 pub mod metadata;
 pub mod non_fungible_resource;
+pub mod non_fungible_resource_with_remote_type;
 pub mod radiswap;
 pub mod transfer_xrd;
