+++ transaction-scenarios/src/scenarios/non_fungible_resource.rs
@@ -1,7 +1,6 @@
 use crate::internal_prelude::*;
 use radix_engine::types::*;
 use radix_engine_interface::api::node_modules::ModuleConfig;
-use radix_engine_interface::blueprints::package::PackageDefinition;
 use radix_engine_interface::*;
 use scrypto::NonFungibleData;
 
@@ -20,8 +19,6 @@ pub struct NonFungibleResourceScenarioState {
     pub integer_non_fungible_resource_with_metadata_standard_data: Option<ResourceAddress>,
     pub integer_non_fungible_resource_with_complex_data: Option<ResourceAddress>,
     pub vault1: Option<InternalAddress>,
-    pub blueprint_with_registered_types: Option<(PackageAddress, String)>,
-    pub non_fungible_resource_using_registered_type: Option<ResourceAddress>,
 }
 
 impl Default for NonFungibleResourceScenarioConfig {
@@ -546,98 +543,6 @@ impl ScenarioCreator for NonFungibleResourceScenarioCreator {
                     )
                 },
             )
-            .successful_transaction_with_result_handler(
-                |core, config, state| {
-                    core.next_transaction_with_faucet_lock_fee(
-                        "non-fungible-resource-create-with-blueprint-type-identifier-prepare",
-                        |builder| {
-                            // Load code and schema
-                            let code = include_bytes!("../../../assets/radiswap.wasm");
-                            let mut schema = manifest_decode::<PackageDefinition>(include_bytes!(
-                                "../../../assets/radiswap.rpd"
-                            ))
-                            .unwrap();
-
-                            // Register `RemoveLiquidityEvent` as TestType
-                            let type_id = if let TypeRef::Static(type_id) = schema.blueprints.values_mut().next().unwrap().schema.events.event_schema.get("RemoveLiquidityEvent").unwrap() {
-                                type_id.clone()
-                            } else {
-                                unreachable!()
-                            };
-                            schema.blueprints.values_mut().next().unwrap().schema.types.type_schema.insert("TestType".to_owned(), type_id);
-
-                            // Build manifest
-                            builder
-                                .get_free_xrd_from_faucet()
-                                .publish_package_advanced(
-                                    None,
-                                    code.to_vec(),
-                                    schema,
-                                    metadata_init! {
-                                        "name" => "Radiswap Package (with type registration)", locked;
-                                    },
-                                    OwnerRole::None,
-                                )
-                                .try_deposit_entire_worktop_or_abort(config.main_account.address, None)
-                        },
-                        vec![],
-                    )
-                },
-                |core, config, state, result| {
-                    state.blueprint_with_registered_types = Some((result.new_package_addresses()[0], "Radiswap".to_owned()));
-                    Ok(())
-                },
-            )
-            .successful_transaction_with_result_handler(
-                |core, config, state| {
-                    core.next_transaction_with_faucet_lock_fee(
-                        "non-fungible-resource-create-with-blueprint-type-identifier",
-                        |builder| {
-                            #[derive(ScryptoSbor, ManifestSbor)]
-                            pub struct RemoveLiquidityEvent {
-                                pub pool_units_amount: Decimal,
-                                pub redeemed_resources: [(ResourceAddress, Decimal); 2],
-                            }
-
-
-                            builder
-                                .call_function(
-                                    RESOURCE_PACKAGE,
-                                    NON_FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT,
-                                    NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT,
-                                    NonFungibleResourceManagerCreateWithInitialSupplyManifestInput {
-                                        owner_role: OwnerRole::None,
-                                        id_type: NonFungibleIdType::Integer,
-                                        track_total_supply: true,
-                                        non_fungible_schema: NonFungibleDataSchema::Remote {
-                                            type_id: BlueprintTypeIdentifier {
-                                                package_address: state.blueprint_with_registered_types.as_ref().unwrap().0,
-                                                blueprint_name: state.blueprint_with_registered_types.as_ref().unwrap().1.clone(),
-                                                type_name: "TestType".to_owned(),
-                                            },
-                                            mutable_fields: index_set_new()
-                                        },
-                                        resource_roles: NonFungibleResourceRoles::single_locked_rule(rule!(allow_all)),
-                                        metadata: metadata! {},
-                                        address_reservation: None,
-                                        entries: indexmap!(
-                                            NonFungibleLocalId::integer(1) => (manifest_decode(&manifest_encode(&RemoveLiquidityEvent {
-                                                pool_units_amount: dec!(5),
-                                                redeemed_resources: [(XRD, dec!(1)), (XRD, dec!(1))]
-                                            }).unwrap()).unwrap(), ),
-                                        ),
-                                    }
-                                )
-                                .try_deposit_entire_worktop_or_abort(config.main_account.address, None)
-                        },
-                        vec![],
-                    )
-                },
-                |core, config, state, result| {
-                    state.non_fungible_resource_using_registered_type = Some(result.new_resource_addresses()[0]);
-                    Ok(())
-                },
-            )
             .finalize(|core, config, state| {
                 Ok(ScenarioOutput {
                     interesting_addresses: DescribedAddresses::new()
@@ -674,14 +579,6 @@ impl ScenarioCreator for NonFungibleResourceScenarioCreator {
                         .add(
                             "integer_non_fungible_resource_with_complex_data",
                             state.integer_non_fungible_resource_with_complex_data.unwrap(),
-                        )
-                        .add(
-                            "package_with_registered_types",
-                            state.blueprint_with_registered_types.clone().unwrap().0,
-                        )
-                        .add(
-                            "non_fungible_resource_using_registered_type",
-                            state.non_fungible_resource_using_registered_type.unwrap(),
                         ),
                 })
             })
